+++ src/mapml/index.js
@@ -54,7 +54,7 @@ import { Util } from './utils/Util';
 import { AttributionButton, attributionButton } from './control/AttributionButton';
 import { ReloadButton, reloadButton } from './control/ReloadButton';
 import { FullscreenButton, fullscreenButton } from './control/FullscreenButton';
-import {attributionControl} from "./control/AttributionControl"; 
+// import {attributionControl} from "./control/AttributionControl"; 
 import {geolocationButton} from "./control/GeolocationButton";
 import { Crosshair, crosshair } from "./layers/Crosshair";
 import { Feature, feature } from "./features/feature";
@@ -662,7 +662,7 @@ M.reloadButton = reloadButton;
 M.FullscreenButton = FullscreenButton;
 M.fullscreenButton = fullscreenButton;
 
-M.attributionControl = attributionControl;
+// M.attributionControl = attributionControl;
 
 M.geolocationButton = geolocationButton;
